move test components into separate folder
