+++ raheem/src/components/TestComponents/Test.js
@@ -1,5 +1,5 @@
 import React, { useState } from "react";
-import firebase from "../firebase";
+import firebase from "../../firebase";
 
 import TestUpdate from "./TestUpdate";
 import TestAdd from "./TestAdd";
