+++ raheem/src/components/TestComponents/TestList.js
@@ -1,5 +1,5 @@
 import React, { useState, useEffect } from "react";
-import firebase from "../firebase";
+import firebase from "../../firebase";
 
 const useItems = () => {
   const [items, setItems] = useState([]); //useState() hook, sets initial state to an empty array
